Unify skeletonize and skeletonize_3d under skeletonize
