+++ skimage/morphology/_skeletonize.py
@@ -2,20 +2,95 @@
 Algorithms for computing the skeleton of a binary image
 """
 
+
 import numpy as np
+from ..util import img_as_ubyte, crop
 from scipy import ndimage as ndi
 
+from .._shared.utils import assert_nD
 from ._skeletonize_cy import (_fast_skeletonize, _skeletonize_loop,
                               _table_lookup_index)
-
-from .._shared.utils import assert_nD
-
-
-# --------- Skeletonization by morphological thinning ---------
+from ._skeletonize_3d_cy import _compute_thin_image
 
 
-def skeletonize(image):
-    """Return the skeleton of a binary image.
+def skeletonize(image, *, method=None):
+    """Compute the skeleton of a binary image.
+
+    Thinning is used to reduce each connected component in a binary image
+    to a single-pixel wide skeleton.
+
+    Parameters
+    ----------
+    image : ndarray, 2D or 3D
+        A binary image containing the objects to be skeletonized. Zeros
+        represent background, nonzero values are foreground.
+    method : {'zhang', 'lee'}, optional
+        Which algorithm to use. Zhang's algorithm [Zha84]_ only works for
+        2D images, and is the default for 2D. Lee's algorithm [Lee94]_
+        works for 2D or 3D images and is the default for 3D.
+
+    Returns
+    -------
+    skeleton : ndarray
+        The thinned image.
+
+    See also
+    --------
+    medial_axis
+
+    References
+    ----------
+    .. [Lee94] T.-C. Lee, R.L. Kashyap and C.-N. Chu, Building skeleton models
+           via 3-D medial surface/axis thinning algorithms.
+           Computer Vision, Graphics, and Image Processing, 56(6):462-478, 1994.
+
+    .. [Zha84] A fast parallel algorithm for thinning digital patterns,
+           T. Y. Zhang and C. Y. Suen, Communications of the ACM,
+           March 1984, Volume 27, Number 3.
+
+
+    Examples
+    --------
+    >>> X, Y = np.ogrid[0:9, 0:9]
+    >>> ellipse = (1./3 * (X - 4)**2 + (Y - 4)**2 < 3**2).astype(np.uint8)
+    >>> ellipse
+    array([[0, 0, 0, 1, 1, 1, 0, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 1, 1, 1, 1, 1, 0, 0],
+           [0, 0, 0, 1, 1, 1, 0, 0, 0]], dtype=uint8)
+    >>> skel = skeletonize(ellipse)
+    >>> skel.astype(np.uint8)
+    array([[0, 0, 0, 0, 0, 0, 0, 0, 0],
+           [0, 0, 0, 0, 0, 0, 0, 0, 0],
+           [0, 0, 0, 0, 0, 0, 0, 0, 0],
+           [0, 0, 0, 0, 1, 0, 0, 0, 0],
+           [0, 0, 0, 0, 1, 0, 0, 0, 0],
+           [0, 0, 0, 0, 1, 0, 0, 0, 0],
+           [0, 0, 0, 0, 1, 0, 0, 0, 0],
+           [0, 0, 0, 0, 0, 0, 0, 0, 0],
+           [0, 0, 0, 0, 0, 0, 0, 0, 0]], dtype=uint8)
+
+    """
+
+    if image.ndim == 2 and (method is None or method == 'zhang'):
+        skeleton = skeletonize_2d(image)
+    elif image.ndim == 3 or (image.ndim == 2 and method == 'lee'):
+        skeleton = skeletonize_3d(image)
+    elif image.ndim == 3 and method == 'zhang':
+        raise ValueError('skeletonize method "zhang" only works for 2D '
+                         'images.')
+    else:
+        raise ValueError('skeletonize requires a 2D or 3D image as input, '
+                         'got {}D.'.format(image.ndim))
+
+
+def skeletonize_2d(image):
+    """Return the skeleton of a 2D binary image.
 
     Thinning is used to reduce each connected component in a binary image
     to a single-pixel wide skeleton.
@@ -498,3 +573,75 @@ def _table_lookup(image, table):
         indexer = _table_lookup_index(np.ascontiguousarray(image, np.uint8))
     image = table[indexer]
     return image
+
+
+def skeletonize_3d(img):
+    """Compute the skeleton of a binary image.
+
+    Thinning is used to reduce each connected component in a binary image
+    to a single-pixel wide skeleton.
+
+    Parameters
+    ----------
+    img : ndarray, 2D or 3D
+        A binary image containing the objects to be skeletonized. Zeros
+        represent background, nonzero values are foreground.
+
+    Returns
+    -------
+    skeleton : ndarray
+        The thinned image.
+
+    See also
+    --------
+    skeletonize, medial_axis
+
+    Notes
+    -----
+    The method of [Lee94]_ uses an octree data structure to examine a 3x3x3
+    neighborhood of a pixel. The algorithm proceeds by iteratively sweeping
+    over the image, and removing pixels at each iteration until the image
+    stops changing. Each iteration consists of two steps: first, a list of
+    candidates for removal is assembled; then pixels from this list are
+    rechecked sequentially, to better preserve connectivity of the image.
+
+    The algorithm this function implements is different from the algorithms
+    used by either `skeletonize` or `medial_axis`, thus for 2D images the
+    results produced by this function are generally different.
+
+    References
+    ----------
+    .. [Lee94] T.-C. Lee, R.L. Kashyap and C.-N. Chu, Building skeleton models
+           via 3-D medial surface/axis thinning algorithms.
+           Computer Vision, Graphics, and Image Processing, 56(6):462-478, 1994.
+
+    """
+    # make sure the image is 3D or 2D
+    if img.ndim < 2 or img.ndim > 3:
+        raise ValueError("skeletonize_3d can only handle 2D or 3D images; "
+                         "got img.ndim = %s instead." % img.ndim)
+
+    img = np.ascontiguousarray(img)
+    img = img_as_ubyte(img, force_copy=False)
+
+    # make an in image 3D and pad it w/ zeros to simplify dealing w/ boundaries
+    # NB: careful here to not clobber the original *and* minimize copying
+    img_o = img
+    if img.ndim == 2:
+        img_o = img[np.newaxis, ...]
+    img_o = np.pad(img_o, pad_width=1, mode='constant')
+
+    # normalize to binary
+    maxval = img_o.max()
+    img_o[img_o != 0] = 1
+
+    # do the computation
+    img_o = np.asarray(_compute_thin_image(img_o))
+
+    # crop it back and restore the original intensity range
+    img_o = crop(img_o, crop_width=1)
+    if img.ndim == 2:
+        img_o = img_o[0]
+    img_o *= maxval
+
+    return img_o
